code refactor - splitting services files
