+++ packages/data-provider/src/core/provider.ts
@@ -14,8 +14,8 @@ import submittedDataRouter from '../routers/submittedDataRouter.js';
 import auditService from '../services/auditService.js';
 import categoryService from '../services/categoryService.js';
 import dictionaryService from '../services/dictionaryService.js';
-import submissionService from '../services/submissionService.js';
-import submittedDataService from '../services/submittedDataService.js';
+import submissionService from '../services/submission/submission.js';
+import submittedDataService from '../services/submittedData/submmittedData.js';
 import * as auditUtils from '../utils/auditUtils.js';
 import * as convertSqonToQueryUtils from '../utils/convertSqonToQuery.js';
 import * as dictionarySchemaRelationUtils from '../utils/dictionarySchemaRelations.js';
